requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0
PyPDF2==3.0.1
python-docx==1.1.0
sentence-transformers==2.2.2
numpy==1.24.3
pydantic==2.5.0
